+++ packages/undercut/src/pull.js
@@ -1,3 +1,4 @@
 export * from "./pull/pull_core.js";
 export * from "./pull/pull_operations.js";
+export * from "./pull/pull_sources.js";
 export * from "./pull/pull_targets.js";