+++ packages/undercut/src/pull/operations/chunk.js
@@ -1,8 +1,7 @@
 import { assert } from "../../utils/assertions.js";
-import { isNumber } from "../../utils/lang.js";
 
 export function chunk(size) {
-	assert(isNumber(size) && size > 0, "Size is required and must be > 0.");
+	assert(Number.isFinite(size) && size > 0, "Size is required and must be > 0.");
 
 	return function* (iterable) {
 		let chunk = [];