+++ packages/undercut/src/pull/operations/nth.js
@@ -1,8 +1,7 @@
 import { assert } from "../../utils/assertions.js";
-import { isNumber } from "../../utils/lang.js";
 
 export function nth(n) {
-	assert(isNumber(n) && n >= 0, "N is required and must be >= 0.");
+	assert(Number.isFinite(n) && n >= 0, "N is required and must be >= 0.");
 
 	return function* (iterable) {
 		let index = 0;